refactor(@angular/cli): split update command into smaller modules
This commit refactors the `ng update` command by extracting
functions from `packages/angular/cli/src/commands/update/cli.ts`
into more focused utility modules:

- `packages/angular/cli/src/commands/update/utilities/git.ts`:
  Encapsulates Git-related operations such as checking for a clean
  repository, creating commits, and retrieving commit hashes.
- `packages/angular/cli/src/commands/update/utilities/cli-version.ts`:
  Handles CLI version compatibility checks, temporary CLI installations,
  and package manager force options.
- `packages/angular/cli/src/commands/update/utilities/constants.ts`:
  Stores shared constants like `ANGULAR_PACKAGES_REGEXP`.
